requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Queue & Caching
celery==5.3.4
redis==5.0.1
flower==2.0.1

# EDI Processing
python-dateutil==2.8.2
pytz==2023.3

# ML
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
# Optional ML enhancements (install separately if needed)
# torch==2.1.0  # For deep learning models (uncomment to enable)
# shap==0.43.0  # For model explainability (uncomment to enable)

# Utilities
structlog==23.2.0
httpx==0.25.2
psutil==5.9.6
urllib3<2.0  # Pin to v1.x for LibreSSL compatibility (macOS)
sentry-sdk[fastapi]==1.38.0

# WebSocket (FastAPI native support, but explicit dependency)
websockets==12.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
black==23.11.0
ruff==0.14.10

